Changes to DB-Structure #10
Also a first upgrade script, that still needs some tests and love

+++ lib/Translator/Installer.php
@@ -32,12 +32,17 @@ class Translator_Installer extends Zikula_AbstractInstaller
      */
     public function install()
     {
+        if (!DBUtil::createTable('translator_moduletranslations')) {
+            return false;
+        }
+        /*
         if (!DBUtil::createTable('translator_modtrans')) {
             return false;
         }
         if (!DBUtil::createTable('translator_translations')) {
             return false;
         }
+        */
         if (!DBUtil::createTable('translator_translations_lang')) {
             return false;
         }
@@ -59,14 +64,16 @@ class Translator_Installer extends Zikula_AbstractInstaller
      */
     public function upgrade($oldversion)
     {
-        /*
         switch ($oldversion) {
             case '1.0.0':
             case '1.0.1':
             case '1.0.2':
             case '1.0.3':
+                if (!$this->to110()) {
+                    return false;
+                }
         }
-        */
+
         // Update successful
         return true;
     }
@@ -78,12 +85,17 @@ class Translator_Installer extends Zikula_AbstractInstaller
      */
     public function uninstall()
     {
+        if (!DBUtil::dropTable('translator_moduletranslations')) {
+            return false;
+        }
+        /*
         if (!DBUtil::dropTable('translator_modtrans')) {
             return false;
         }
         if (!DBUtil::dropTable('translator_translations')) {
             return false;
         }
+        */
         if (!DBUtil::dropTable('translator_translations_lang')) {
             return false;
         }
@@ -95,6 +107,69 @@ class Translator_Installer extends Zikula_AbstractInstaller
         $this->delVars();
 
         // Deletion successful
+        return true;
+    }
+
+    /**
+     * Upgrade routine to Version 1.1.0
+     *
+     * @return boolean
+     */
+    private function to110()
+    {
+        if (!DBUtil::createTable('translator_moduletranslations')) {
+            return false;
+        }
+
+        DBUtil::executeSQL("ALTER TABLE translator_moduletranslations
+            ADD COLUMN transmod_id_old INT UNSIGNED NULL AFTER ignore,
+            ADD COLUMN trans_id_old INT UNSIGNED NULL AFTER transmod_id_old");
+        DBUtil::executeSQL("ALTER TABLE translator_occurrences RENAME TO translator_occurrences_old");
+        DBUtil::createTable('translator_occurrences');
+        DBUtil::executeSQL("ALTER TABLE translator_translations_lang RENAME TO translator_translations_lang_old");
+        DBUtil::createTable('translator_translations_lang');
+        $res = DBUtil::executeSQL("select a.transmod_id, a.trans_id, a.mod_id, b.sourcestring
+            from translator_modtrans a
+                left join translator_translations b
+                    on a.trans_id = b.trans_id
+            where a.in_use = 1");
+        $rows = DBUtil::marshallObjects($res);
+
+        foreach ($rows as $row) {
+            DBUtil::executeSQL("insert into translator_moduletranslations
+                (module_id, sourcestring, in_use, ignore, transmod_id_old, trans_id_old)
+                values
+                ({$row['mod_id']}, '{$row['sourcestring']}', 1, 0, {$row['transmod_id']}, {$row['trans_id']})");
+        }
+
+        $res = DBUtil::executeSQL("select * from translator_moduletranslations");
+        $rows = DBUtil::marshallObjects($res);
+
+        foreach ($rows as $row) {
+            $res1 = DBUtil::executeSQL("select * from translator_occurrences_old where transmod_id = {$row['transmod_id_old']}");
+            $rows1 = DBUtil::marshallObjects($res1);
+            foreach ($rows1 as $r1) {
+                DBUtil::executeSQL("insert into translator_occurrences
+                    (transmod_id, file, line)
+                    values
+                    ({$row['trans_id']}, '{$r1['file']}', {$r1['line']})");
+            }
+            $res2 = DBUtil::executeSQL("select * from translator_translations_lang_old where trans_id = {$row['trans_id_old']}");
+            $rows2 = DBUtil::marshallObjects($res2);
+            foreach ($rows2 as $r2) {
+                DBUtil::executeSQL("insert into translator_translations_lang
+                    (trans_id, language, targetstring)
+                    values
+                    ({$row['trans_id']}, '{$r2['language']}', '{$r2['targetstring']}')");
+            }
+        }
+
+        DBUtil::executeSQL("ALTER TABLE translator_moduletranslations DROP COLUMN trans_id_old, DROP COLUMN transmod_id_old");
+        DBUtil::dropTable('translator_occurrences_old');
+        DBUtil::dropTable('translator_translations_lang_old');
+        DBUtil::dropTable('translator_modtrans');
+        DBUtil::dropTable('translator_translations');
+
         return true;
     }
 }
